requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
psycopg2-binary
slowapi
python-dotenv
pyjwt
passlib
bcrypt==4.0.1
python-multipart
email-validator
pandas
numpy
requests
beautifulsoup4
lxml
aiohttp
aiolimiter
aiofiles
fake_useragent